Converted all files to typscript equivalent
