+++ code/beastly-brawl-showdown/.meteor/packages
@@ -13,7 +13,7 @@ standard-minifier-css@1.9.3   # CSS minifier run for production mode
 standard-minifier-js@3.0.0    # JS minifier run for production mode
 es5-shim@4.8.1                # ECMAScript 5 compatibility for older browsers
 ecmascript@0.16.10              # Enable ECMAScript2015+ syntax in app code
-typescript@5.6.3              # Enable TypeScript syntax in .ts and .tsx modules
+typescript              # Enable TypeScript syntax in .ts and .tsx modules
 shell-server@0.6.1            # Server-side component of the `meteor shell` command
 hot-module-replacement@0.5.4  # Update client in development without reloading the page
