+++ code/beastly-brawl-showdown/client/main.tsx
@@ -5,6 +5,6 @@ import { App } from '/imports/ui/App';
 
 Meteor.startup(() => {
   const container = document.getElementById('react-target');
-  const root = createRoot(container);
+  const root = createRoot(container!);
   root.render(<App />);
 });
