+++ code/beastly-brawl-showdown/imports/api/links.ts
@@ -1,0 +1,10 @@
+import { Mongo } from 'meteor/mongo';
+
+export interface Link {
+  _id?: string;
+  title: string;
+  url: string;
+  createdAt: Date;
+}
+
+export const LinksCollection = new Mongo.Collection<Link>('links');
