+++ code/beastly-brawl-showdown/server/main.ts
@@ -1,7 +1,7 @@
 import { Meteor } from 'meteor/meteor';
-import { LinksCollection } from '/imports/api/links';
+import { Link, LinksCollection } from '/imports/api/links';
 
-async function insertLink({ title, url }) {
+async function insertLink({ title, url }: Pick<Link, 'title' | 'url'>) {
   await LinksCollection.insertAsync({ title, url, createdAt: new Date() });
 }
