+++ code/beastly-brawl-showdown/tests/main.ts
@@ -1,0 +1,21 @@
+import { Meteor } from 'meteor/meteor';
+import assert from 'assert';
+
+describe('meteor-app', function () {
+  it('package.json has correct name', async function () {
+    const { name } = await import('../package.json');
+    assert.strictEqual(name, 'meteor-app');
+  });
+
+  if (Meteor.isClient) {
+    it('client is not server', function () {
+      assert.strictEqual(Meteor.isServer, false);
+    });
+  }
+
+  if (Meteor.isServer) {
+    it('server is not client', function () {
+      assert.strictEqual(Meteor.isClient, false);
+    });
+  }
+});
